COMMERCE-2372 talend - component - definition - Reorganize - move duplicated code to base class
